requirements.txt
fastapi
uvicorn
openai
fireworks-ai
google-generativeai
python-dotenv
